Compatibility with MSVC

+++ anyoption.cpp
@@ -992,7 +992,7 @@ AnyOption::readFile( const char* fname )
 		return NULL;
 	}
 	is.seekg (0, ios::end);
-	length = is.tellg();
+	length = (int) is.tellg();
 	is.seekg (0, ios::beg);
 	//changed by Hugo
 	//previously  buffer = (char*) malloc(length*sizeof(char));
